chore(shell-api): combine common Cursor + AbstractCursor code
Create an `AbstractCursor` class (named like the driver equivalent)
as a common base class for `Cursor` and `AggregationCursor`
to reduce code duplication.
