+++ packages/shell-api/src/cursor.ts
@@ -4,59 +4,36 @@ import {
   returnsPromise,
   returnType,
   serverVersions,
-  ShellApiClass,
   shellApiClassDefault,
-  toShellResult,
   deprecated
 } from './decorators';
 import {
   ServerVersions,
-  asPrintable,
   CURSOR_FLAGS
 } from './enums';
-import {
+import type {
   FindCursor as ServiceProviderCursor,
   CursorFlag,
   Document,
   CollationOptions,
-  ExplainVerbosityLike,
   ReadPreferenceLike,
   ReadConcernLevelId,
   TagSet,
   HedgeOptions
 } from '@mongosh/service-provider-core';
-import { iterate, validateExplainableVerbosity, markAsExplainOutput } from './helpers';
 import Mongo from './mongo';
-import { CursorIterationResult } from './result';
 import { printWarning } from './deprecation-warning';
+import { AbstractCursor } from './abstract-cursor';
 
 @shellApiClassDefault
 @hasAsyncChild
-export default class Cursor extends ShellApiClass {
-  _mongo: Mongo;
+export default class Cursor extends AbstractCursor {
   _cursor: ServiceProviderCursor;
-  _currentIterationResult: CursorIterationResult | null = null;
   _tailable = false;
-  _batchSize: number | null = null;
 
   constructor(mongo: Mongo, cursor: ServiceProviderCursor) {
-    super();
+    super(mongo);
     this._cursor = cursor;
-    this._mongo = mongo;
-  }
-
-  /**
-   * Internal method to determine what is printed for this class.
-   */
-  async [asPrintable](): Promise<CursorIterationResult> {
-    return (await toShellResult(this._currentIterationResult ?? await this._it())).printable;
-  }
-
-  async _it(): Promise<CursorIterationResult> {
-    const results = this._currentIterationResult = new CursorIterationResult();
-    await iterate(results, this._cursor, this._batchSize ?? await this._mongo._batchSize());
-    results.cursorHasMore = !this.isExhausted();
-    return results;
   }
 
   /**
@@ -100,15 +77,10 @@ export default class Cursor extends ShellApiClass {
   }
 
   @returnType('Cursor')
-  batchSize(size: number): Cursor {
-    this._batchSize = size;
+  batchSize(size: number): this {
+    super.batchSize(size);
     this._cursor.batchSize(size);
     return this;
-  }
-
-  @returnsPromise
-  async close(options: Document): Promise<void> {
-    await this._cursor.close(options);
   }
 
   @returnType('Cursor')
@@ -132,42 +104,6 @@ export default class Cursor extends ShellApiClass {
   }
 
   @returnsPromise
-  async explain(verbosity?: ExplainVerbosityLike): Promise<any> {
-    // TODO: @maurizio we should probably move this in the Explain class?
-    // NOTE: the node driver always returns the full explain plan
-    // for Cursor and the queryPlanner explain for AggregationCursor.
-    if (verbosity !== undefined) {
-      verbosity = validateExplainableVerbosity(verbosity);
-    }
-    const fullExplain: any = await this._cursor.explain(verbosity);
-
-    const explain: any = {
-      ...fullExplain
-    };
-
-    if (
-      verbosity !== 'executionStats' &&
-      verbosity !== 'allPlansExecution' &&
-      explain.executionStats
-    ) {
-      delete explain.executionStats;
-    }
-
-    if (verbosity === 'executionStats' &&
-      explain.executionStats &&
-      explain.executionStats.allPlansExecution) {
-      delete explain.executionStats.allPlansExecution;
-    }
-
-    return markAsExplainOutput(explain);
-  }
-
-  @returnsPromise
-  async forEach(f: (doc: Document) => void): Promise<void> {
-    return this._cursor.forEach(f);
-  }
-
-  @returnsPromise
   async hasNext(): Promise<boolean> {
     if (this._tailable) {
       printWarning(
@@ -176,42 +112,13 @@ export default class Cursor extends ShellApiClass {
         this._mongo._internalState.context.print
       );
     }
-    return this._cursor.hasNext();
-  }
-
-  @returnsPromise
-  async tryNext(): Promise<Document | null> {
-    return this._cursor.tryNext();
-  }
-
-  async* [Symbol.asyncIterator]() {
-    let doc;
-    while ((doc = await this.tryNext()) !== null) {
-      yield doc;
-    }
+    return super.hasNext();
   }
 
   @returnType('Cursor')
   hint(index: string): Cursor {
     this._cursor.hint(index);
     return this;
-  }
-
-  isClosed(): boolean {
-    return this._cursor.closed;
-  }
-
-  isExhausted(): boolean {
-    return this.isClosed() && this.objsLeftInBatch() === 0;
-  }
-
-  @returnsPromise
-  async itcount(): Promise<number> {
-    let count = 0;
-    while (await this.tryNext()) {
-      count++;
-    }
-    return count;
   }
 
   @returnType('Cursor')
@@ -221,9 +128,8 @@ export default class Cursor extends ShellApiClass {
   }
 
   @returnType('Cursor')
-  map(f: (doc: Document) => Document): Cursor {
-    this._cursor.map(f);
-    return this;
+  map(f: (doc: Document) => Document): this {
+    return super.map(f);
   }
 
   @returnType('Cursor')
@@ -233,9 +139,8 @@ export default class Cursor extends ShellApiClass {
   }
 
   @returnType('Cursor')
-  maxTimeMS(value: number): Cursor {
-    this._cursor.maxTimeMS(value);
-    return this;
+  maxTimeMS(value: number): this {
+    return super.maxTimeMS(value);
   }
 
   @returnType('Cursor')
@@ -260,7 +165,7 @@ export default class Cursor extends ShellApiClass {
         this._mongo._internalState.context.print
       );
     }
-    return this._cursor.next();
+    return super.next();
   }
 
   @returnType('Cursor')
@@ -276,9 +181,8 @@ export default class Cursor extends ShellApiClass {
   }
 
   @returnType('Cursor')
-  projection(spec: Document): Cursor {
-    this._cursor.project(spec);
-    return this;
+  projection(spec: Document): this {
+    return super.projection(spec);
   }
 
   @returnType('Cursor')
@@ -312,15 +216,13 @@ export default class Cursor extends ShellApiClass {
   }
 
   @returnType('Cursor')
-  skip(value: number): Cursor {
-    this._cursor.skip(value);
-    return this;
+  skip(value: number): this {
+    return super.skip(value);
   }
 
   @returnType('Cursor')
-  sort(spec: Document): Cursor {
-    this._cursor.sort(spec);
-    return this;
+  sort(spec: Document): this {
+    return super.sort(spec);
   }
 
   @returnType('Cursor')
@@ -340,7 +242,7 @@ export default class Cursor extends ShellApiClass {
   }
 
   @returnType('Cursor')
-  pretty(): Cursor {
+  pretty(): this {
     return this;
   }
 
@@ -357,10 +259,6 @@ export default class Cursor extends ShellApiClass {
   showRecordId(): Cursor {
     this._cursor.showRecordId(true);
     return this;
-  }
-
-  objsLeftInBatch(): number {
-    return this._cursor.bufferedCount();
   }
 
   @returnType('Cursor')
